[REFACTOR]Seperate structs of each types
