+++ drawingApp/drawingApp/Model/Alpha.swift
@@ -1,0 +1,23 @@
+//
+//  Alpha.swift
+//  drawingApp
+//
+//  Created by Kai Kim on 2022/03/02.
+//
+
+import Foundation
+enum Alpha : Double, CaseIterable, CustomStringConvertible {
+    case One = 0.1
+    case Two
+    case Three
+    case Four
+    case Five
+    case Six
+    case Seven
+    case Eight
+    case Nine
+    case Ten
+    var description: String {
+        return "Alpha \()"
+    }
+}
